+++ src/FileException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Forestry\Log;
+
+class FileException extends \RuntimeException
+{
+}